feat(frontend): implement housing view mobilization tab
